Updated files to silence warnings about NSInteger being 64-bit. (Comment out a log of NSLog statements here) Added functions to check for iOS version to constants.h. Fixed main.m not knowing which class was the app delegate. Fixed application window not having a root controller. Fix for iOS8: Added prompt to ask user for permission to display notifications. (TODO: Add similar notification for MapKit)

+++ Classes/PickerViewController.m
@@ -97,7 +97,7 @@
 - (IBAction)save:(id)sender
 {
 	NSInteger row = [customPickerView selectedRowInComponent:0];
-	[delegate didPickPurpose:row];
+	[delegate didPickPurpose: (int)row];
 }
